Uninstalls testing-library/react-hooks as its been combined into testing-library/react
On a clean pull of the repository, you were unable to run `npm install` because of conflicting peer dependencies with the react version installed, and what @testing-library/react-hooks expected. Turns out for React 18, @testing-library/react-hooks was combined into @testing-library/react. So we can uninstall it. It was not being used so we just needed to uninstall.